refact: Move key splitting into a module & make into an async generator
Note that the b-tree walking in attempt to get the ranges for the
threads is not functional at all. It only walks in one direction so is
lopsided and only covers the initial range because the end is appended
onto the ranges creating a range that likely has half the keys.

The async generator pattern for looping through the split up key space
allows support for async operations to get keys (such as the theoretical
b-tree walking).