feat: Extend Hyperbee adding parallelReadStream as a new method
Extending Hyperbee means a few things are simplier in the code. First
options passed to create the hyperbee are saved to pass to the worker
threads. Second, creating a pool on starting a parallel read is now more
manageable as the pool's scope is only for that read. Finally it will
help support locking the hyperbee for reads in the future.

Moved simple test / example into its own script to showcase how the
`parallelReadStream()` can run and how it performs.