chore: Pack example keys w/ `lexicographic-integer` to make worst case
Because `lexicographic-integer` packs the integers for the keys by
lengthening the buffer and prepending larger numbers with a 'high'
prefix aka more and more `f`s, this skews the key space towards one end
making the division key splitter no longer effective as virtually all
keys (anything greater than ~200 it seems) land in the last bin.

The b-tree solution for key space splitting looks to be the best
solution moving forward if completed correctly.

Finally noted that the `unlocked: true` wasn't necessary currently. This
is likely because the example very clearly waits for the read to happen
and so isn't trying to write at the same time.