feat: Split Range based on b-tree in hyperbee
This implementation seems to be more consistently even due to the
b-trees self balancing nature. It is not however guaranteed to be even
with my naive tree walking algorithm. This likely can be improved but
there seems to be a trade off of speed vs exactness as further walking
the tree will require more blocks be read up front that eats into the
purpose of the multithreading optimization.

There is likely bugs & flaws in the tree walking algorithm as I haven't
accounted for when a key is below the _gKey (aka lexicographically lower
key) which could have children that include values above the key.