chore: Add artificial 'complex' value simulation for evaluating decoding
This was to create a setup to test if decoding the node value in the
worker was more or less efficient than sending the node not decoded to
the main thread and having it decode it there.

Initial tests with transferable buffers seemed to be slightly faster
when decoded on the main thread, but not significantly.